refactor: move key-specific database features into local trait
