+++ src/keys/mod.rs
@@ -1,4 +1,7 @@
-mod router;
+pub mod database;
+pub mod entities;
 pub mod service;
+
+mod router;
 
 pub use router::router;
